requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.5.0
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3
pyobjc-framework-Cocoa==10.3
pyobjc-framework-Quartz==10.3
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rubicon-objc==0.4.9
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
ultralytics==8.2.28
ultralytics-thop==0.2.7
urllib3==2.2.1
